+++ test/warning_test.py
@@ -13,6 +13,14 @@ True
 >>> g_.data2 == g.data2
 True
 >>> with open('data.json', 'w') as fp:
+...     json.dump({'data1': [0], 'data2': [0], '__version__': '0'}, fp)
+>>> with warns(RuntimeWarning, match=r"Source matplotlib version(.*) or Python version(.*) is different from yours, plots may be different."):
+...     g = GenerativeImage(data=open('data.json', 'r'))
+>>> with open('config.json', 'w') as fp:
+...     json.dump({'f1': 'x', 'f2': 'y', '__version__': '0'}, fp)
+>>> with warns(RuntimeWarning, match=r"Source matplotlib version(.*) or Python version(.*) is different from yours, plots may be different."):
+...     g = GenerativeImage(config=open('config.json', 'r'))
+>>> with open('data.json', 'w') as fp:
 ...     json.dump({'data1': [0], 'data2': [0], 'python_version': '0'}, fp)
 >>> with warns(RuntimeWarning, match=r"Source matplotlib version(.*) or Python version(.*) is different from yours, plots may be different."):
 ...     g = GenerativeImage(data=open('data.json', 'r'))